move 4 files to the login file in interface_adapter
