+++ src/interface_adapter/Login/LoginState.java
@@ -1,0 +1,4 @@
+package interface_adapter.Login;
+
+public class LoginState {
+}
